feat: Final spike implementation amendments. Refactoring, styling and efficiency.
