+++ src/api/instances.tsx
@@ -21,6 +21,7 @@ import { addEntitlements } from "util/entitlements/api";
 import { addTarget } from "util/target";
 import { linkForInstanceDetail } from "util/instances";
 import { ROOT_PATH } from "util/rootPath";
+import type { LxdFileExplorerItem } from "types/fileExplorer";
 
 export const instanceEntitlements = [
   "can_access_console",
@@ -551,4 +552,49 @@ export const createInstanceBackup = async (
     .then((data: LxdOperationResponse) => {
       return data;
     });
+};
+
+export const fetchInstanceFile = async (
+  name: string,
+  project: string,
+  path: string,
+): Promise<LxdFileExplorerItem> => {
+  const params = new URLSearchParams();
+  params.set("project", project);
+  params.set("path", path);
+
+  return fetch(
+    `${ROOT_PATH}/1.0/instances/${encodeURIComponent(name)}/files?${params.toString()}`,
+    {
+      method: "GET",
+    },
+  )
+    .then(handleResponse)
+    .then((data: LxdFileExplorerItem) => {
+      return data;
+    });
+};
+
+export const fetchInstanceFileHeader = async (
+  name: string,
+  project: string,
+  path: string,
+): Promise<Response> => {
+  const params = new URLSearchParams();
+  params.set("project", project);
+  params.set("path", path);
+
+  return fetch(
+    `${ROOT_PATH}/1.0/instances/${encodeURIComponent(name)}/files?${params.toString()}`,
+    {
+      method: "HEAD",
+    },
+  ).then((response: Response) => {
+    if (!response.ok) {
+      throw new Error(
+        `Failed to fetch instance file header: ${response.status} ${response.statusText}`,
+      );
+    }
+    return response;
+  });
 };
