+++ src/sass/_file_explorer.scss
@@ -1,0 +1,5 @@
+.fe-item {
+  align-items: center;
+  display: flex;
+  gap: 0.5rem;
+}
