+++ src/sass/styles.scss
@@ -98,6 +98,7 @@ $border-thin: 1px solid $colors--theme--border-low-contrast !default;
 @import "device_detail_list";
 @import "empty_state";
 @import "error_page";
+@import "file_explorer";
 @import "file_row";
 @import "form-link";
 @import "font";
